Move the UI support for storage initizalization to pyanaconda.ui.lib
Move the functions reset_storage, reset_bootloader and select_all_disks_by_default
to pyanaconda.ui.lib.
