+++ pyanaconda/storage/initialization.py
@@ -21,22 +21,13 @@ from gi.repository import BlockDev as blockdev
 
 from blivet import util as blivet_util, udev, arch
 from blivet.devicelibs import crypto
-from blivet.errors import StorageError
 from blivet.flags import flags as blivet_flags
 from blivet.static_data import luks_data
 
 from pyanaconda.anaconda_logging import program_log_lock
 from pyanaconda.core.configuration.anaconda import conf
-from pyanaconda.core.constants import BOOTLOADER_DRIVE_UNSET
-from pyanaconda.errors import errorHandler as error_handler, ERROR_RAISE
-from pyanaconda.modules.common.constants.objects import DISK_SELECTION, BOOTLOADER, DEVICE_TREE
-from pyanaconda.modules.common.constants.services import STORAGE
-from pyanaconda.modules.common.task import sync_run_task
 from pyanaconda.storage.osinstall import InstallerStorage
 from pyanaconda.platform import platform
-
-from pyanaconda.anaconda_loggers import get_module_logger
-log = get_module_logger(__name__)
 
 
 def enable_installer_mode():
@@ -111,71 +102,3 @@ def load_plugin_s390():
     plugin.name = blockdev.Plugin.S390
     plugin.so_name = None
     blockdev.reinit([plugin], reload=False)
-
-
-def reset_storage(scan_all=False, retry=True):
-    """Reset the storage model.
-
-    :param scan_all: should we scan all devices in the system?
-    :param retry: should we allow to retry the reset?
-    """
-    # Clear the exclusive disks to scan all devices in the system.
-    if scan_all:
-        disk_select_proxy = STORAGE.get_proxy(DISK_SELECTION)
-        disk_select_proxy.SetExclusiveDisks([])
-
-    # Scan the devices.
-    storage_proxy = STORAGE.get_proxy()
-
-    while True:
-        try:
-            task_path = storage_proxy.ScanDevicesWithTask()
-            task_proxy = STORAGE.get_proxy(task_path)
-            sync_run_task(task_proxy)
-        except StorageError as e:
-            # Is the retry allowed?
-            if not retry:
-                raise
-            # Does the user want to retry?
-            elif error_handler.cb(e) == ERROR_RAISE:
-                raise
-            # Retry the storage reset.
-            else:
-                continue
-        else:
-            # No need to retry.
-            break
-
-    # Reset the partitioning.
-    storage_proxy.ResetPartitioning()
-
-
-def reset_bootloader():
-    """Reset the bootloader."""
-    bootloader_proxy = STORAGE.get_proxy(BOOTLOADER)
-    bootloader_proxy.SetDrive(BOOTLOADER_DRIVE_UNSET)
-
-
-def select_all_disks_by_default():
-    """Select all disks for the partitioning by default.
-
-    It will select all disks for the partitioning if there are
-    no disks selected. Kickstart uses all the disks by default.
-
-    :return: a list of selected disks
-    """
-    disk_select_proxy = STORAGE.get_proxy(DISK_SELECTION)
-    selected_disks = disk_select_proxy.SelectedDisks
-    ignored_disks = disk_select_proxy.IgnoredDisks
-
-    if not selected_disks:
-        # Get all disks.
-        device_tree = STORAGE.get_proxy(DEVICE_TREE)
-        all_disks = device_tree.GetDisks()
-
-        # Select all disks.
-        selected_disks = [d for d in all_disks if d not in ignored_disks]
-        disk_select_proxy.SetSelectedDisks(selected_disks)
-        log.debug("Selecting all disks by default: %s", ",".join(selected_disks))
-
-    return selected_disks
