+++ pyanaconda/ui/lib/storage.py
@@ -15,10 +15,14 @@
 # License and may only be used or replicated with the express permission of
 # Red Hat, Inc.
 #
+from blivet.errors import StorageError
 
 from pyanaconda.anaconda_loggers import get_module_logger
-from pyanaconda.core.constants import PARTITIONING_METHOD_AUTOMATIC
+from pyanaconda.core.constants import PARTITIONING_METHOD_AUTOMATIC, BOOTLOADER_DRIVE_UNSET
+from pyanaconda.errors import errorHandler as error_handler, ERROR_RAISE
+from pyanaconda.modules.common.constants.objects import DISK_SELECTION, BOOTLOADER, DEVICE_TREE
 from pyanaconda.modules.common.constants.services import STORAGE
+from pyanaconda.modules.common.task import sync_run_task
 
 log = get_module_logger(__name__)
 
@@ -41,3 +45,71 @@ def find_partitioning():
         )
 
     return STORAGE.get_proxy(object_path)
+
+
+def reset_storage(scan_all=False, retry=True):
+    """Reset the storage model.
+
+    :param scan_all: should we scan all devices in the system?
+    :param retry: should we allow to retry the reset?
+    """
+    # Clear the exclusive disks to scan all devices in the system.
+    if scan_all:
+        disk_select_proxy = STORAGE.get_proxy(DISK_SELECTION)
+        disk_select_proxy.SetExclusiveDisks([])
+
+    # Scan the devices.
+    storage_proxy = STORAGE.get_proxy()
+
+    while True:
+        try:
+            task_path = storage_proxy.ScanDevicesWithTask()
+            task_proxy = STORAGE.get_proxy(task_path)
+            sync_run_task(task_proxy)
+        except StorageError as e:
+            # Is the retry allowed?
+            if not retry:
+                raise
+            # Does the user want to retry?
+            elif error_handler.cb(e) == ERROR_RAISE:
+                raise
+            # Retry the storage reset.
+            else:
+                continue
+        else:
+            # No need to retry.
+            break
+
+    # Reset the partitioning.
+    storage_proxy.ResetPartitioning()
+
+
+def reset_bootloader():
+    """Reset the bootloader."""
+    bootloader_proxy = STORAGE.get_proxy(BOOTLOADER)
+    bootloader_proxy.SetDrive(BOOTLOADER_DRIVE_UNSET)
+
+
+def select_all_disks_by_default():
+    """Select all disks for the partitioning by default.
+
+    It will select all disks for the partitioning if there are
+    no disks selected. Kickstart uses all the disks by default.
+
+    :return: a list of selected disks
+    """
+    disk_select_proxy = STORAGE.get_proxy(DISK_SELECTION)
+    selected_disks = disk_select_proxy.SelectedDisks
+    ignored_disks = disk_select_proxy.IgnoredDisks
+
+    if not selected_disks:
+        # Get all disks.
+        device_tree = STORAGE.get_proxy(DEVICE_TREE)
+        all_disks = device_tree.GetDisks()
+
+        # Select all disks.
+        selected_disks = [d for d in all_disks if d not in ignored_disks]
+        disk_select_proxy.SetSelectedDisks(selected_disks)
+        log.debug("Selecting all disks by default: %s", ",".join(selected_disks))
+
+    return selected_disks
